Add signal level to Domoticz controller
Added signal level to the Domoticz controller commands.
The RSSI domain was empiricaly maped to signal level from 0 to 11.
Values below -95dBm report a signal level  of 0 and values above -35dBm
correspond to a level of 11.
I chose those values based on the ESP not showing surrounding APs with
RSSI lower than -95dBm and the maximum value of about -25dBm when the
ESP was sitting on the router.